requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.3
apscheduler>=3.10.0
